requirements.txt
fastapi==0.115.0
psycopg2-binary==2.9.9
yt-dlp==2024.9.27
youtube-transcript-api==0.6.2
anthropic==0.34.2
pydantic==2.9.2
